restructure folders for ngrx
