+++ src/app/components/cart-add/cart-add.component.ts
@@ -1,9 +1,6 @@
-import { ThisReceiver } from '@angular/compiler';
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { Observable } from 'rxjs';
-import { Products } from 'src/app/models/Products';
-import { ProductServiceService } from 'src/app/services/product-service.service';
+import { ProductServiceService } from '../../services/product/product-service.service';
 
 @Component({
   selector: 'cart-add',
@@ -18,16 +15,16 @@ export class CartAddComponent implements OnInit {
   constructor(productService: ProductServiceService, private route: ActivatedRoute) {this.productService = productService}
 
   ngOnInit(): void {
-   
+
     let id = parseInt(this.route.snapshot.paramMap.get('id')!);
     this.productId = id;
-    
+
     this.productService.getProductById(this.productId).subscribe((product) => {
      this.product = product;
    })
 
-   
-    
+
+
     console.log(this.product)
   }
 }
