+++ src/app/components/cart-checkout/cart-checkout.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
 import { Products } from 'src/app/models/Products';
-import { CartCheckoutService } from 'src/app/services/cart-checkout.service';
+import { CartCheckoutService } from '../../services/cart/cart-checkout.service';
 
 @Component({
   selector: 'app-cart-checkout',
@@ -12,7 +11,7 @@ export class CartCheckoutComponent implements OnInit {
 
 
   cart:Products[]=[];
-  
+
   constructor(private cartCheckoutService: CartCheckoutService) { }
 
   ngOnInit(): void {
@@ -21,7 +20,7 @@ export class CartCheckoutComponent implements OnInit {
   }
 
   addProductsToCart(){
-    
+
     this.cartCheckoutService.addProductsToCart()
     .subscribe(products=>{
       products.forEach(product=>this.cart.push(product))
@@ -32,7 +31,7 @@ export class CartCheckoutComponent implements OnInit {
     // const idx = this.cart.findIndex(product=>product==item);
 
     // if(idx > -1) {
-    //   this.cart[idx] = 
+    //   this.cart[idx] =
 
     // }
   }
