+++ src/app/services/cart/cart-checkout.service.spec.ts
@@ -1,3 +1,4 @@
+import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing';
 
 import { CartCheckoutService } from './cart-checkout.service';
@@ -6,7 +7,9 @@ describe('CartCheckoutService', () => {
   let service: CartCheckoutService;
 
   beforeEach(() => {
-    TestBed.configureTestingModule({});
+    TestBed.configureTestingModule({
+      imports: [ HttpClientTestingModule ],
+    });
     service = TestBed.inject(CartCheckoutService);
   });
