+++ src/app/services/product/product-service.service.ts
@@ -1,9 +1,9 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { Products } from '../models/Products';
-import { User } from '../models/User';
-import { UserRole } from '../models/UserRole';
+import { Products } from '../../models/Products';
+import { User } from '../../models/User';
+import { UserRole } from '../../models/UserRole';
 
 @Injectable({
   providedIn: 'root'
